% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jerichojars-package.R
\docType{package}
\name{jerichojars}
\alias{jerichojars}
\alias{jerichojars-package}
\title{Contents of the 'Jericho HTML Parser' Java archive by
Martin Jericho \url{http://jericho.htmlparser.net/docs/index.html} provided to
support functions in the 'jericho' package.}
\description{
Contents of the 'Jericho HTML Parser' Java archive by
Martin Jericho \url{http://jericho.htmlparser.net/docs/index.html} provided to
support functions in the 'jericho' package.
}
\author{
Bob Rudis (bob@rud.is)
}
